circle cutting operational, demo recorded

+++ image_recorder.py
@@ -63,7 +63,7 @@ def start_listening(interval=.01):
     while True:
 
         now = rospy.get_rostime()
-        now = now.secs + now.nsecs/1e9
+        now = int(now.secs + now.nsecs/1e9)
 
         left = imgsub.left_image
         right = imgsub.right_image
